+++ src/escuadra/modulos/matematicas/conversor_longitud_extendido.py
@@ -1,4 +1,4 @@
-from src.escuadra.modulos.matematicas.conversor_longitud import UNIDADES_A_METROS as UNIDADES_BASE
+from escuadra.modulos.matematicas.conversor_longitud import UNIDADES_A_METROS as UNIDADES_BASE
 
 UNIDADES_A_METROS = {
     **UNIDADES_BASE,